chore: shift to global setup
