+++ apps/studio/package.json
@@ -181,7 +181,7 @@
     "@vitest/coverage-istanbul": "^2.1.2",
     "autoprefixer": "^10.4.20",
     "clean-webpack-plugin": "^4.0.0",
-    "dotenv-cli": "^7.3.0",
+    "dotenv-cli": "^7.4.4",
     "eslint": "^9.10.0",
     "k6-trpc": "^1.0.0",
     "mockdate": "^3.0.5",
